attach a page->category dict to all scrapers
show it in the viewer and export to elasticsearch

that was a nasty task and the result is not entirely straight.
The page index might change over time and i tried not to be
too exact, but at times i might have been.. the things that
should fit for all channels are news, weather and sport.
the rest is experimental.